Naive update

See merge request aging-team/aging!4

+++ src/gameoflife/Game.h
@@ -3,18 +3,20 @@
 
 #include "CellMap.h"
 
+#include <cstddef>
+
 class Game
 {
 public:
-	Game();
-	~Game();
-	void drawCell(unsigned int x, unsigned int y, int alive);
-	void drawGrid();
-	void nextGeneration();
+  Game();
+  ~Game();
+  void drawCell(size_t x, size_t y, int alive);
+  void drawGrid();
+  void nextGeneration();
 
 private:
-	CellMap* map;
-	unsigned long generation;
+  CellMap* map;
+  unsigned long generation;
 };
 
 #endif // __GAME_H
